test: move getCIDR function to internal/util and add test code
move to util/internal and add getCIDR function test file
to test the getCIDR function and added test code for the part
that checks the format of the subnetmask
